utils: centralize utility functions in dedicated module to avoid duplication
