+++ reverse-proxy/src/handler/common/mod.rs
@@ -1,4 +1,3 @@
 pub(crate) mod consts;
-pub mod utils;
 pub mod types;
 pub mod handler;
