+++ utils/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "utils"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+reqwest = { version = "0.11", features = ["json"] }
+uuid = { version = "1.16.0", features = ["v4"] }
+serde_json = "1.0.140"
+serde = { version = "1.0.219", features = ["derive"] }
